Model Changes. Used external csproj for models
