+++ Graph-Ting.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33502.453
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Graph-Ting", "Graph-Ting\Graph-Ting.csproj", "{083475A6-7219-4DF4-B5B0-A91CE352CFC9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Graph-Ting", "Graph-Ting\Graph-Ting.csproj", "{083475A6-7219-4DF4-B5B0-A91CE352CFC9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphTing.Models", "GraphTing.Models\GraphTing.Models.csproj", "{2E972646-2B2A-4449-AEFD-1F5AA4302C60}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{083475A6-7219-4DF4-B5B0-A91CE352CFC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{083475A6-7219-4DF4-B5B0-A91CE352CFC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{083475A6-7219-4DF4-B5B0-A91CE352CFC9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2E972646-2B2A-4449-AEFD-1F5AA4302C60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2E972646-2B2A-4449-AEFD-1F5AA4302C60}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2E972646-2B2A-4449-AEFD-1F5AA4302C60}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2E972646-2B2A-4449-AEFD-1F5AA4302C60}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
