+++ Graph-Ting/Graph-Ting.csproj
@@ -8,4 +8,8 @@
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\GraphTing.Models\GraphTing.Models.csproj" />
+  </ItemGroup>
+
 </Project>
