+++ Graph-Ting/MainWindow.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Windows;
 using System.Windows.Controls;
+using GraphTing.Models.BinaryTrees;
 
 namespace Graph_Ting
 {
